career_management/requirements.txt
django
djangorestframework
django-jazzmin
django-cors-headers
whitenoise
django-extensions
Pillow
djangorestframework-simplejwt